Feat: Introducing incident tab in table details page
