+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/de-de.json
@@ -575,6 +575,7 @@
     "inactive-announcement-plural": "Inaktive Ankündigungen",
     "incident": "Incident",
     "incident-manager": "Incident Manager",
+    "incident-plural": "Incidents",
     "incident-status": "Incident Status",
     "include": "Einschließen",
     "include-entity": "{{entity}} einschließen",
