+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -575,6 +575,7 @@
     "inactive-announcement-plural": "Inactive Announcements",
     "incident": "Incident",
     "incident-manager": "Incident Manager",
+    "incident-plural": "Incidents",
     "incident-status": "Incident Status",
     "include": "Include",
     "include-entity": "Include {{entity}}",
