+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/he-he.json
@@ -575,6 +575,7 @@
     "inactive-announcement-plural": "הכרזות לא פעילות",
     "incident": "Incident",
     "incident-manager": "Incident Manager",
+    "incident-plural": "Incidents",
     "incident-status": "Incident Status",
     "include": "כלול",
     "include-entity": "כלול {{entity}}",
