+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/nl-nl.json
@@ -575,6 +575,7 @@
     "inactive-announcement-plural": "Inactieve aankondigingen",
     "incident": "Incident",
     "incident-manager": "Incidentmanager",
+    "incident-plural": "Incidents",
     "incident-status": "Incidentstatus",
     "include": "Inclusief",
     "include-entity": "{{entity}} opnemen",
