+++ openmetadata-ui/src/main/resources/ui/src/pages/TableDetailsPageV1/TableDetailsPageV1.tsx
@@ -40,6 +40,7 @@ import SchemaTab from '../../components/Database/SchemaTab/SchemaTab.component';
 import TableQueries from '../../components/Database/TableQueries/TableQueries';
 import { QueryVote } from '../../components/Database/TableQueries/TableQueries.interface';
 import EntityRightPanel from '../../components/Entity/EntityRightPanel/EntityRightPanel';
+import IncidentManager from '../../components/IncidentManager/IncidentManager.component';
 import Lineage from '../../components/Lineage/Lineage.component';
 import { EntityName } from '../../components/Modals/EntityNameModal/EntityNameModal.interface';
 import PageLayoutV1 from '../../components/PageLayoutV1/PageLayoutV1';
@@ -742,6 +743,26 @@ const TableDetailsPageV1: React.FC = () => {
               table={tableDetails}
               testCaseSummary={testCaseSummary}
             />
+          ),
+      },
+      {
+        label: (
+          <TabsLabel
+            id={EntityTabs.INCIDENT}
+            name={t('label.incident-plural')}
+          />
+        ),
+        key: EntityTabs.INCIDENT,
+        children:
+          tablePermissions.ViewAll || tablePermissions.ViewTests ? (
+            <div className="p-x-lg p-b-lg p-t-md">
+              <IncidentManager
+                isIncidentPage={false}
+                tableDetails={tableDetails}
+              />
+            </div>
+          ) : (
+            <ErrorPlaceHolder type={ERROR_PLACEHOLDER_TYPE.PERMISSION} />
           ),
       },
       {
